restructuration of components and validators #6
