+++ web/src/app/component/auth/register/register.component.scss
@@ -64,6 +64,6 @@
     color: red;
     justify-content: left;
     flex-direction: column;
-    gap: 1px; /* Ajoute un espace entre chaque élément */
+    gap: 5px; /* Ajoute un espace entre chaque élément */
   }
